+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - `cli/list`: Show plugin braces only there's plugin output
 - `lib/tman`: Change return value of function tman_deinit()
+- `lib/unit`: Change return value of function unit_delpgn()
 
 ### Fixed